Add particle_number target (#478)
* add clustering ground truth

* temp training

* stability fixes

* refactor OC loss

* OC aggregation

* fixes for OC

* add matching summary and unit test

* fix(data): ensure consistent particle_number assignment and validate CMS ground truth

- Added 'particle_number' to Y_FEATURES in CMS and EDM4HEP dataset utils to resolve shape mismatches.
- Improved 'particle_number' assignment in CMS postprocessing by propagating original CaloParticle IDs through the graph.
- Added automated ground truth validation for CMS in 'tests/validate_cms_gt.py'.
- Optimized local test scripts for CMS and CLD by reducing evaluation samples and using 'uv run'.
- Cleaned up miscellaneous formatting and unused imports in losses and postprocessing scripts.

* chore: bump all dataset versions to 3.2.0 to reflect Y_FEATURES shape change

* readd other samples

* assign hits inclusively

* fixes for tesla

* fix test

* add missing files

* fix pre-commit

* limit to ttbar

* fix gt

* test: add detailed specifications to test suite and automate data fetching

- Added detailed technical docstrings (Specs) to all 25 files in the 'tests/' directory, explaining what functionalities they validate and identifying dependencies (like mock data or external files).
- Created 'scripts/fetch_test_data_cld.sh' and 'scripts/fetch_test_data_cms.sh' to centralize and automate the downloading and post-processing of external ROOT files into Parquet/Pickle formats needed for validation.
- Refactored 'scripts/local_test_cld.sh' and 'scripts/local_test_cms.sh' to use the new dedicated fetch scripts, reducing redundancy and improving script robustness with pushd/popd navigation.

* add OC logging

* pre-commit

* fix dtypes, docs

* enable OC layers only when the OC loss is enabled

* fix preds

* detach beta

* update OC loss, allow sampler to run from scratch

* fix(model): mask null particle predictions/targets and improve object condensation loss

- In object condensation (OC) loss, refactored regression targets to be hit-level rather than aggregating them per-particle. Now identifying representative hits (with cls_id > 0) to broadcast canonical target properties (pt, energy, eta, sin_phi, cos_phi) to all member hits of a particle.
- Compute correct log-scale physical target values relative to the individual hit baseline, zero out losses for noise hits, and scale pt/energy losses by sqrt of target physical pt.
- Replace aggregated PID classification loss in OC loss with hit-level classification loss.
- Zero out predicted/target values (pt, energy, eta, sin_phi, cos_phi, phi, p4, momentum) in standard/OC loss, inference predictions, and evaluation when the ground truth cls_id == 0 (null particles).
- Correct cluster-level beta-weighted average calculations in MLPF predictions by squeezing the weights array to avoid shape mismatches.
- Update evaluation to print validation table for diagnostic visibility, add log-level event quality metrics (sum_pt, sum_e, particle counts), and handle dataset splitting edge cases.
- Add test coverage for zeroing out regression components when no target particle is present.

* chore: strip object condensation loss, predictions, LossType config, and OC_V/OC_beta keys while keeping particle_number target

* fix test

* enable other datasets

* cleanup some remaining OC stuff

* cleanup

+++ mlpf/plotting/draw_graphs.py
@@ -21,7 +21,7 @@ def main(args):
 
     features = ["eta", "phi"]
 
-    x_data = data.x.cpu().detach().numpy()
+    x_data = data.x.cpu().detach().float().numpy()
 
     mask = (x_data[:, 4] == 0) & (x_data[:, 5] == 0) & (x_data[:, 6] == 0) & (x_data[:, 7] == 0)
     # good_index = np.zeros((x_data.shape[0], 1, 2), dtype=int)
@@ -35,7 +35,7 @@ def main(args):
     df["isTrack"] = ~mask
 
     row, col = data.edge_index.cpu().detach().numpy()
-    y_truth = data.y.cpu().detach().numpy()
+    y_truth = data.y.cpu().detach().float().numpy()
 
     input_dim = data.x.shape[1]
     hidden_dim = 32
